vmm: added vm_control module
Added parts of the vm_control crate from crosvm as a module in the vmm
crate. The contents of this module currently consist of the VmRequest
and VmResponse enums, which will be used when new mmio devices are added
by the device manager, and will be even more useful when we also
implement control sockets, if we choose to follow the crosvm model.

Signed-off-by: Alexandru Agache <aagch@amazon.com>