firecracker: added virtio blk and net dev support
The firecracker app now supports additional parameters that allow one
block device and/or one net device to be attached to the VM. Block
devices are backed by the specified file. Net devices communicate with
the host via a TAP device that is created with the specified ip address
and subnet mask (the subnet mask defaults to /24).

Also added code for attaching blk and net virtio devices based on the
machine configuration, and for responding to MmioRead and MmioWrite
cpu exit events.

Signed-off-by: Alexandru Agache <aagch@amazon.com>