vmm: added EpollContext struct
Added the EpollContext struct to vmm/src/lib.rs, which will be used to
handle epoll events, especically in the context of a single control
thread for all devices. It can be extended to allow multiplexing
control logic over multiple threads.

The EpollContext is already used to take over the control loop, but no
devices are currently added (they still use their own threads). At this
point, the run_control() method effectively deals only with EXIT or
SERIAL related events (same as before), but uses an EpollContext
object to do so.

Signed-off-by: Alexandru Agache <aagch@amazon.com>