virtio/queue.rs: added helper logic for tests
Added a couple of helper structs and their implementation to
virtio/queue.rs, in the devices crate. These objects can be used to
easily access parts of a VIRTIO queue directly in the underlying memory,
(as opposed to looking at the state of structs such as Queue, and going
forward from there), which is useful to see if various operations have
the expected result, or to setup different testing scenarios.

Signed-off-by: Alexandru Agache <aagch@amazon.com>