vmm/src/lib.rs: separate attach block devices from boot_kernel
Added block_devices_configs to the Vmm struct.
block_devices_config is a struct defined in device_configs that
contains a list of Block Devices that need to be attached
before booting the kernel.

In boot_kernel, if there is at least one root device added,
we will attached them all. The devices should be added as part
of the api call PUT to /devices/{device_id}. For now, we add
a dummy device from the command line.

Signed-off-by: Andreea Florescu <fandree@amazon.com>