Change EpollContext to allow removal of EventFDs from the epoll loop
This patch enables clean/safe removal of events from EpollContext.

The kernel epoll underlying mechanism is given 'tokens' which end up
indexing specific data structures based on the activated event.
Just removing the eventfd from the kernel epoll is not enough to
completely disable that event and safely remove the associated data
structure because at the time of 'EPOLL_CTL_DEL', the kernel epoll
may have already 'dispatched' that event and the vmm event loop has
not yet handled it (event handling is serialized in the vmm event loop).
That means we need to take extra precautions and also mark events as
removed in the EpollContext so that any in-flight events are not handled
post-removal.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>