vm config: add structure for configuring the VM
Added VirtualMachineConfiguration structure that contains
the configurable parameters of the virtual machine (nr of vcpu
and memory size).
Added a function for setting the VM configuration that is used
from main and will also be used to setup the machine from the API.
Removed vcpu_count and mem_size from MachineCfg.

Signed-off-by: Andreea Florescu <fandree@amazon.com>