net_util/tap: added open_named() method
Added the open_named() method to the Tap implementation. This allows
the Firecracker process to open an already existing TAP interface
identified by its device name. Since opening a TUN/TAP is similar to
creating one, if the specfied device does not exist, and Firecracker is
run with sufficient privileges, a new device with that particular name
will be created.

Signed-off-by: Alexandru Agache <aagch@amazon.com>