main: attach net iface via cmdline parameter
After we added functionality to attach network devices via API calls,
doing the same using Firecracker cmdline parameters was no longer
possible. This commit brings back support for this feature, based on
some of the recently added logic. Moreover, the cmdline parameter for
adding network devices now expects the name of an already existing
TAP interface (which will be temporarily created for the lifetime of the
process if it doesn't actually exist, and we run with sufficient
privileges).

Signed-off-by: Alexandru Agache <aagch@amazon.com>