Virtio block: implemented read-only behaviour
Added logic to properly treat the presence of the is_read_only PUT block
device request. Read-only block devices open their backing files as
read-only. If the root device is also read-only, the the "ro" option is
added to the kernel cmdline. Also, added the --read-only-root parameter
to the vmm-no-api subcommand.

Signed-off-by: Alexandru Agache <aagch@amazon.com>