small refactor of get_device_handler() for code clarity
Using a match on the Option and then doing get_or_insert() on it
is redundant and confusing. Ideally get_or_insert() would directly
be used, but that would mean calling channel recv() before each
check which is also undesirable.

This patch verbosely sets the value and then returns a mutable
reference to it.

While the code proposed by this patch is more verbose, it is also
clearer in intent and usage.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>