virtio rate limiter implementation
Add a rate limiter implementation tailored to virtio usage.
The RateLimiter uses two token buckets to limit rate on bytes/s and ops/s.

Token accounting is done as much as possible as part of the normal data flow,
meaning both consumption and replenishing of tokens is done during the normal
virtio virtqueue processing code flow.
Only when budget is over and virtio queue processing is blocked
When budget is over, virtio queue processing is blocked and a timer is scheduled
to unblock it after a while (this period will need to be fine-tuned later).
The timer mechanism used is timer_fd which will be integrated in the VMM event loop.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>