cpuid: filter cpu features based on template
For now, the only existing template is based on the
T2.medium instance type and it is named T2.
The template is set as part of the vCPU configure.

Setting the CPU Template includes 2 steps:
* set the cpu basic information (model, family et. al)
* disable features that are not available on T2

KVM masks out the cpu model string, so as a first iteration,
the CPU model string is set to Intel(R) Xeon(R) Processor.

Signed-off-by: Andreea Florescu <fandree@amazon.com>