vmm: added check for required KVM capabilities
This commit contains a couple of changes which were necesssary in order
to add checks for required KVM capabilities before trying to actually
start a VM. It introduces the KvmContext struct, which now provides the
Kvm struct when required by rest of the code, and can only be used as
such after all checks successfully pass. It also contains other relevant
data which can be obtained during the initialization process, such as the
maximum number of vcpus supported by KVM (other things will probably be
added in the future).

Signed-off-by: Alexandru Agache <aagch@amazon.com>