virtio net: make sure we process all packets from TAP
Fix a highly sporadic bug where the last packets meant for the guest
would never be sent from the TAP to the guest by Firecracker.

BUG breakdown:
Let's assume that at some point, the guest is sent through the TAP
a burst of N packets followed by nothing, thus generating a single
RX_TAP_EVENT in Firecracker. In a scenario where at that moment
virtio-net rx queue only has available capacity for M < N-1 packets
what happens is:
 1. Firecracker handles RX_TAP_EVENT and starts pulling packets
    from tap and pushing them to the guest virtio-net rx queue.
 2. The (M+1)th packet will be read from the tap but rx_single_frame()
    will fail since there is no more room. This results in breaking
    the process_rx() loop and setting 'deferred_rx' to true.
 3. After a while the guest will consume some of the rx_queue packets
    and available buffers will be made available in virtio-net rx_queue.
    On this event, Firecracker gets the RX_QUEUE_EVENT and checks
    'deferred_rx', find it to be true and handles the deferred (M+1)th
    packet.
 4. If there is no other new RX_TAP_EVENT to restart the loop, packets
    [M+2, N] will just sit queued on the tap, but Firecracker will not
    grab and push them to the guest.

The fix:
On RX_QUEUE_EVENT, if there was at least one deferred frame, try and push
it to the guest and if it succeeds check the tap for pending frames and
process them if available.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>