cpuid: make all cores part of one socket
We don't gracefully handle the topology in which we have
more than one socket. For multiple sockets, we would need to
change the cache topology (L3 is shared at package level). This
also complicates the logic for the case where hyperthreading is
disabled because we shouldn't have packages with 1 core and
packages with 2 cores in the same topology.

Signed-off-by: Andreea Florescu <fandree@amazon.com>