jailer: Add initial implementation of the firecracker jailer
At the moment, the application isn't compatible with firecracker yet. It
simply takes an id and an exec file. When jailing firecracker, the
expectation is that the id will be the instance id while the exec file
is an absolute path to the firecracker executable.

The id will later be used to create the "environment". The environment
is made up of the cgroup plus a run directory containing the API socket
- which will be created before hand - and the chroot.

Signed-off-by: Filippo Sironi <sironi@amazon.de>