jailer: Add support for initializing and using cgroup
Create a cgroup hierarchy (e.g.,
/sys/fs/cgroup/firecracker/i-0c00b0ecbe29a6520) to further jail the
current process with it.
The reason for having a hierarchy is that it provides customers with a
easy way to limit the resources available to a certain exec file (e.g.,
firecracker) en-mass.

Use the resources from the root cgroup to initialize the hierarchy for
now. This will probably change in the future since we may get
information to constrain an instance of exec file to a certain socket,
for example.

Signed-off-by: Filippo Sironi <sironi@amazon.de>