vmm/lib.rs: event loop resumes on EINTR
Since epoll::wait() is a blocking operation, it was possible for it to
return EINTR after the reception of a signal that did not cause the
process to be killed. This should be an extremely rare occurence (in
fact, we don't expect Firecracker to receive any such signals), but
we now handle the return of EINTR from epoll::wait() by restarting
the loop (as customary in many similar situations).

Signed-off-by: Alexandru Agache <aagch@amazon.com>