logger/lib.rs: internal logger changes
Refactored and reworked the definition of the Logger struct to allow the
type to become Sync and have internal mutability. The end goal here is
to have a static global Logger variable which can be used both for
logging and writing metrics (seen as a special operation, which is not
triggered via the log macros).

Signed-off-by: Alexandru Agache <aagch@amazon.com>