api: updated put for machine config
Added the cpu template to MachineConfiguration struct.
The default value for the cpu_template is none thus firecracker
starts without a template.

The function for configuring the vcpu needs most of the fields
from the MachineConfiguration (except mem_size_mib), so instead of
passing them individually, just use the whole struct.

Signed-off-by: Andreea Florescu <fandree@amazon.com>