logger: added log_helper method
Added the log_helper method to the Logger implementation. Currently,
it expects a msg: String parameter. In a future PR we'll update the
way things are written to the selected destination to avoid the creation
and allocation of unnecessary intermediate Strings. The log_helper
method takes care of the common logic involved in both writing regular
log messages, and dumping metrics.

Signed-off-by: Alexandru Agache <aagch@amazon.com>