cpuid: fix bug in T2 Template
When trying to disable the CPU features from leaf 0x1,
register ecx, we were actually setting the ebx register.
This is why we couldn't disable the sdbg flag (IA32_DEBUG_INTERFACE MSR
for silicon debug). When trying to disable this flag, FC
would crash because we would actually set the cache line size in bytes.

Signed-off-by: Andreea Florescu <fandree@amazon.com>