firecracker: added jailer stub
Added jailer.rs to src/bin. This is currently a stub, but will evolve
into the actual jailer binary. By placing it in src/bin (src being the
folder which contains the Firecracker main.rs), it will get
automatically built and placed in the target directory, alongisde the
Firecracker binary.

Signed-off-by: Alexandru Agache <aagch@amazon.com>