cpuid: check if hyperthreading is available on host
... before enabling/disabling it.

When the host has logical cpus as hyperthreads, but the client
disables hyperhtreads through the API, we will panic and firecracker
should crash after logging an error. The reason behind panic-ing
is that on the host the hyperthreads of the same core are not
co-scheduled. We want to make sure that the client is aware that
hyperthreading on the host cannot be disabled by Firecracker, but
it has to be disabled before running Firecracker with ht_enabled
set to false.

When the host has logical cpus as full cores, but the client
enables hyperthreading, we will log a warning and continue with the
execution because there is no security issue involved here, but rather
the topology of the guest will report full cores as hyperhtreads.

Signed-off-by: Andreea Florescu <fandree@amazon.com>