jailer: properly remove FD_CLOEXEC
We know first read the fd flags, unset the FD_CLOEXEC bit, and only then
call fcntl with F_SETFD and the resulting flags. This is better than the
old approach where we just set the fd flags to 0, which would have been
fine only for as long as FD_CLOEXEC remains the only valid flag.

Signed-off-by: Alexandru Agache <aagch@amazon.com>