jailer: removed JailerArg struct
This struct was essentially used to convert the string arguments received
by the jailer binaries into typed values which were then passed to the
rest of the program logic. However, it turns out having this as an
intermediate hop of sorts when interpreting the parameters became quite
awkward for a larger number of parameters, and due to borrowing
restrictions in Rust. What we do now is we pass the ArgMatches struct to
the Env::new() constructor, and handle all validation there.

Signed-off-by: Alexandru Agache <aagch@amazon.com>