jailer: fixed cgroup config inheritance
We had a potential race condition, which could arise if a jailer process
created a .../firecracker/<id> cgroup subfolder, but did not get to
populate cgroup config files such as cpuset.mems. Now, when we attempt to
inherit the value in such a file from the parent (the "firecracker"
folder), we repeat the process for the grand parent whenever an empty
line is encountered.

Signed-off-by: Alexandru Agache <aagch@amazon.com>