Use sigaction instead of signal
The `signal` function is not thread safe and should not be used in
multithreaded processes, according to the man page. The correct way
would be to use the `sa_handler` field in the `sigaction` structure to
set the same handler that used to be set via `signal`; however, the
implementation of `struct sigaction` in the `libc` crate doesn't have
that field, so the handler needs to take a `siginfo` struct as
parameter, whether or not it needs one.

Fixes #351

Signed-off-by: Alexandra Iordache <aghecen@amazon.com>