kvm: fix nasty bug triggering high cpu usage
Channel 2 of the PIT is used to control the PC speaker.
Not emulating one would trigger a vm exit in userspace
everytime when the 0x61 (i.e. KVM_SPEAKER_BASE_ADDRESS)
address was written to.
This commit enables the PC speaker by setting the
KVM_PIT_SPEAKER_DUMMY flag when creating the PIT.

Signed-off-by: Diana Popa <dpopa@amazon.com>