tests: added unique ip generator class
The IP generator class is meant to be used as a singleton. It
should not be instantiated directly, but rather the static function
get_instance should be used. The UniqueIPv4Generator generates IPs
that are locally administrated and receives as argument the
length of the subnet mask. The subnet mask cannot be changed when
invoking get_instance meaning that the submask will be the same
across all tests in a test session.

The class has two methods to be used by test fixtures:
- get_next_available_ips(count) -> returns a list of unique IPs
  across the current test session. It only returns valid IPs of a
  subnet (first and last IPs of a subnet are not returned as they
  are reserved)
- get_next_available_subnet_range -> returns a tuple with 2 IPs.
  The first IP is the subnet ID and the last IP is the broadcast
  IP of the subnet.

These two methods are thread safe.

Signed-off-by: Andreea Florescu <fandree@amazon.com>