tests: added network_config fixture
The fixture is to be used in testing the microvm image that have
the SSH capability. This fixture yields a UniqueIPv4Generator
that can be used for setting unique IPs for tap devices and microvms.

Added a default value for the subnet mask as it is going to be used
when initializing the UniqueIPv4Generator as well as for the setting
the subnet of the tap device.

Added another CPU topology test that uses ssh to test that the
IPs are indeed unique.

Signed-off-by: Andreea Florescu <fandree@amazon.com>