tests: add retry when initializing the ssh conn
It might take some time for the network interface of the VM to
be configured. If the VM didn't got the time to configure the net
interface, the SSH Connection might fail with the error
NoValidConnectionsError. Connect to the VM using retry_call which
allows the function to be called 5 times with a 5 seconds delay
before throwing a connection exception.

Signed-off-by: Andreea Florescu <fandree@amazon.com>