jailer: moved /dev/net/tun creation logic
Move the logic which created the /dev/net folder within the jail, and
the one which used mknod to create the actual device to different parts
of the source code. Also, the test from jailer/src/uuid.rs is now part
of a tests module, as the ones from other files.

Signed-off-by: Alexandru Agache <aagch@amazon.com>