tests: refactored basic config
Added the root file system configuration to the basic config.
Renamed the function _test_default_block_devices to put_default_scratch_device
and added it to the microvm class as it can be used by other tests that are not
part of test_api.py.

Used basic config for cpu tests.

Signed-off-by: Andreea Florescu <fandree@amazon.com>