virtio-net: prepare tx scatter-gather IO
We're currently using an intermediate buffer to build the TX packet from
the virtqueue descriptors.
Ideally, we should use scatter-gather and avoid using this intermediate
buffer altogether.

This commit does part of the setup needed for scatter-gather by recording
the addrs and lens of the descriptors and postponing the memcpy to right
before writing to the tap.

This commit still uses the intermediate buffer, but slightly helps with
keeping the cache hot when doing the actual memcpy().

This change also greatly improves cpu usage when a TX rate-limiter is present.
Before, the full packet was built and copied over to the intermediate buffer,
but if there wasn't enough rate-limiter budget for it, that work would just be
wasted.
With this patch, the packet length is determined while recording the iovec,
and if there is not enough rate-limiter budget, the function returns fast
without copying over any memory.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>