loader: add support for bzImage kernels.
Tries loading as ELF first, then falls back to bzImage.
bzImage is slower to boot than ELF, because the kernel has to
decompress itself.

Added an init_size field to the bootparams block, which the
decompressor uses to know how much space it has to work with.  Made it
big enough for a moderate sized kernel.  We really ought to be extracting
the bootparams block from the kernel instead.

Signed-off-by: Tim Deegan <tdeegan@amazon.com>