jailer: added --daemonize parameter
The --daemonize parameter makes the jailer call setsid() and redirect
the three stdio descriptors to /dev/null right before exec-ing into the
target binary. This commit also modifies the logic in enable_stdin_event
from vmm/src/lib.rs, because the stdin event cannot be added to epoll
when redirected to /dev/null (or so it seems).

Signed-off-by: Alexandru Agache <aagch@amazon.com>