tests: use a separate netns for each MicrovmSlot
Each MicrovmSlot is now associated with its own network namespace. This
is further passed on to the jailer, which spawns the Microvm in that
namespace. For now, commands which involve the network stack run by
temporarily switching to the network namespace of the slot (either
via 'ip netns exec', or using the nsenter module). At some point, we may
want to create proper routing between the default network namespace
(where the jailer currently runs), and those associated with each slot.

Signed-off-by: Alexandru Agache <aagch@amazon.com>