vmm: add instrumentation for measuring guest boot time
A timestamp is taken on receiving the `StartInstance` API request.

Guest boot-time is the time passed from `StartInstance` to running
the guest userspace `init`. To measure this time we depend on the
guest `init`, or another process, to write a magic value to a magic
IO port.

The magic IO port used is 0x03f0 which is the first io port address
in the range reserved for the Floppy Disk Controller. This port was
chosen since we do not support or plan on supporting a Floppy Disk
Controller.

The magic value of 123 was pseudo-randomly chosen.

On every write of 123 to the 0x3f0 io port, the time difference
between the io-write trap and the StartInstance timestamp is
printed in the Firecracker log at the warn level using format:
"Guest-boot-time = {:>6} us {} ms"

Signed-off-by: Adrian Catangiu <acatan@amazon.com>