micro_http: added support for HTTP/1.1
We now check in the request line if the HTTP version is either 1.0 or 1.1.

For parsing the request URI, we have to also accept URIs as absolute URIs
or relative URIs, where the relative URI is an absolute path.

Added get method for http_version in Request. This is to be used when
building a Response for a Request. If the request parsing failed, the

Response will have the default HTTP Version (HTTP/1.1). Exported the Version
struct publicly so it can be used when building a response.

Signed-off-by: Andreea Florescu <fandree@amazon.com>