micro_http: implemented write instead of raw
For the response to be sent to the client, it needs to be converted
to a u8 array. For this purpose, we were using Vec<u8> and kept
allocating memory without actually needing it. With this refactoring,
all parts of the request are used as bytes and written directly to
a buffer.

Signed-off-by: Andreea Florescu <fandree@amazon.com>