Serialize machine-config fields to correct datatypes
This change corrects the GET /machine-config API to encode the
MachineConfiguration values using the same JSON datatypes that
Firecracker expects for input.

Although the MachineConfiguration struct has defaults for its value,
there is no contract enforcing that this will remain the case in the
future. So we take care to provide our own defaults in case we're
trying to serialize a completely uninitialized instance.

Signed-off-by: Noah Meyerhans <nmeyerha@amazon.com>