fc_util: added fn read_cycles()
This function returns a timestamp expressed in CPU cycles (pretty much
what rdtsc() does on x86). It can be used for accurate and cheap
relative time measurements. There might be some synchronization issues
if a Firecracker is allowed to jump from one socket to another, and
there will most likely be synchronization problems if we migrate the
Firecracker process from one host to another, but these scenarios do
not seem likely given our current use cases. Moreover, we could rethink
this function as something that returns an opaque (ideally monotonically
increasing) value, if/when we start caring about all those things. We
have to judge on a case by case basis whether this makes sense in the
parts of the code where it's going to appear. Anyway it's currently
implemented for x86_64 only, based on rdtsc().

Signed-off-by: Alexandru Agache <aagch@amazon.com>