vmm: gracefully shut down on reboot
When the guest sends the reboot command, we call the stop function.
Before shutting down, we flush the metrics and we clean up the
terminal (setting canonical mode and disabling the stdin event).

As we don't support shutting down an instance and then restarting it,
there is no reason for keeping the API thread and the VMM thread running.
On stop, also exit from the Firecracker process.

Also, removed the parameter api_enabled from run_control() as we only
use the API for starting the VMM for now.

Signed-off-by: Andreea Florescu <fandree@amazon.com>