data_model: removed FIRECRACKER_IS_JAILED
Removed the aforementioned global variable and reworked existing logic
to accomodate this change. When Firecracker is jailed, some parameters
are propagated from main.rs to the appropriate places in the
implementation of struct Kvm, and that of ApiServer, in order to
distinguish between having to create resources, and reusing a fd already
opened by the jailer.

Signed-off-by: Alexandru Agache <aagch@amazon.com>