vmm & api_server refactoring
The api_server now depends on the vmm and imports all
the necessary structures from it. The structures are defined
in the vmm_config crate.

Redefined the one shot channels by which the vmm and the API
communicate so that on each channel we put bits of information
that are in a vmm specific language.

As much as possible, renamed the structures in the VMM so they
do not have words related to the API in them.

Refactored the error handling. We now have 2 types of errors:
* errors associated with the VMM public interface (VmmActionError).
  Each of the sub-categories of this error kind corresponds to an
  action from the public interface as defined in the enum
  VmmAction.
* errors associated with the internal state of the VMM or its
  components. These errors cannot be directly triggered by the user.

Signed-off-by: Andreea Florescu <fandree@amazon.com>