renamed kernel_loader to kernel
The plan is to add all kernel related functionality to this module
as separate modules. One of the first things that will be moved here
is the Kernel Command Line.
The first module in this new crate is the loader that contains all
the code that was previously in the kernel_loader crate.

Signed-off-by: Andreea Florescu <fandree@amazon.com>