vmm: removed unnecessary pubs
The public interface of the VMM should contain just the following
structures, enums and types along with their implementation:
* VmmAction & VmmActionError
* the oneshot channels: OutcomeSender & OutcomeReceiver
* the data sent on the channels: VmmRequestOutcome
and the function start_vmm_thread, which is called from the main
crate.
Removed all the other pubs.

Moved test code to mod tests and added the test directive where
it was missing.

Signed-off-by: Andreea Florescu <fandree@amazon.com>