vmm: removed NetworkInterfaceBody
Previously, we had a NetworkInterfaceConfig that contained the
NetworkInterfaceBody. The only additional field of the
NetworkInterfaceConfig was the tap device.

Renamed the NetworkInterfaceBody to NetworkInterfaceConfig.

Added the tap field to NetworkInterfaceConfig. This field cannot
and should not be serialized or deserialized because it is created
using the field `host_dev_name`.

Reworked the update and create of the NetworkInterfaceConfig. Added
validate functions for both actions. Also, checked if the host_dev_name
is in use before trying to open the Tap, so we can return a meaningful
error in that case.

Signed-off-by: Andreea Florescu <fandree@amazon.com>