api_server: fixed bugs in parse_drives_req
For PATCH requests we were not actually checking that the id from
the path matches the id from the body. Modified into_parsed_request
to also validate the id.

Removed Parsed Request Dummy that was used for GET requests.

Added unit tests for all error cases and assert that the error is
what we are expecting.

Signed-off-by: Andreea Florescu <fandree@amazon.com>