dumbo/tcp: added NextSegmentStatus logic
The NextSegmentStatus enum represents whether a particular Connection
or Endpoint has an immediate segment to send, will have one if a
particular timeout expires, or simply does not have anything. Together
with the associated logic which returns this status, this enables the
TcpHandler to return results quicker when its endpoints are idle.

Signed-off-by: Alexandru Agache <aagch@amazon.com>