dumbo/ns: improved performance of write_next_frame
Changed the way the MMDS ns attempts to send the next frame out of the
TcpHandler. Whereas before we ended up iterating through every existing
connection looking for something to send, we now use a couple of new
helper methods to determine much quicker if there's an outgoing packet
available.

Signed-off-by: Alexandru Agache <aagch@amazon.com>