kvm: removed constructors of VcpuFd and VmFd
As per the KVM API, the VM fd is created using the KVM fd.
Similarly, the vcpu fd is created using the VM fd. As the kvm
crate offers wrappers over KVM ioctls, we should try to keep the
public interface of this crate as close as possible to the KVM
api.

Signed-off-by: Andreea Florescu <fandree@amazon.com>