tests: don't daemonize jailer by default
This is a workaround for #554. Sometimes, the cloned process hangs. It
remains stuck in a sequence of FUTEX_WAIT (which fail with ETIMEDOUT)
and FUTEX_WAKE (which don't wake anyone). Could be a deadlock in the
Python implementation, or something else. In the meantime, to unblock
the CI suite, running the jailer in a screen session, undaemonized,
seems to work.

Signed-off-by: Alexandra Iordache <aghecen@amazon.com>