devtool: Fix test_shut_down.py hang
The Docker container was being launched without a default init process
to reap all zombie processes inside the container. This caused the
firecracker process to hang around in a defunct state, which, in turn,
caused test_shut_down.py to timeout, waiting for the firecracker PID to
disappear.
Fix: now launching Docker with a default init process (`docker run
--init`).

Signed-off-by: Dan Horobeanu <dhr@amazon.com>