dumbo: some fixes to handler and ns
Fixed a bug in the handler logic which caused some RST segments to be
sent without filling in the appropriate destination. Updated tests to
check for this condition. Also did some slight changes to the logic
responsible for sending ARP replies in ns. We now check if the reply
can be sent successfully first before discarding it.

Signed-off-by: Alexandru Agache <aagch@amazon.com>