dumbo: other small fixes to handler and ns
Added a new error condition to the TCP handler, which triggers if the IP
destination address of an incoming packet is invalid. Modified the tests
to also take this into account. Fixed a small issue in ns, where the
number of destroyed connections was not incremented for an EndpointDone
event signalled by the inner handler.

Signed-off-by: Alexandru Agache <aagch@amazon.com>