docs: Adjust tenets and project role description
We are always gathering new information, and therefore project tenets
are not set in stone. We want to thoughtfully adopt tenets that help
us make tough calls so they should not change often. But we want to
make sure that it is known that the door is always open to refine our
thinking.

For this reason, we traditionally include "unless you know better
ones" when we state our tenets.

Signed-off-by: Matt Wilson <msw@amazon.com>