jailer: drop un-needed chroot(2)
There is no practical difference before and after preforming a chroot
into "/" -- the VFS cannot differentiate between the two cases
(current_chrooted will actually return false because the root of the
namespace is the same as current->fs). It should also be noted that
chroot(2) provides practically no protection when compared to
pivot_root(2).

However, it is technically necessary for us to chdir("/") because
pivot_root(2) doesn't guarantee that your cwd will be changed correctly
-- so switching explicitly is required (doing chroot("/") acted as a
more complicated chdir("/")).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>