CI: run unit tests sequentially
By default, `cargo test` runs the tests in parallel. Official
recommended solution is to use environment variable `RUN_TEST_THREADS`
to specify the number of threads used to run the tests. Since we are
opening `/dev/kvm` in multiple unit tests, we need this in order to
avoid race conditions.

Signed-off-by: Diana Popa <dpopa@amazon.com>