Apply seccomp filters to all firecracker threads
Refactor and move the seccomp setup to before creating the vCPUs,
that way the vCPU threads inherit the filters from the VMM thread.

Also pass a closure that sets up the seccomp filters to the API
thread to be called after the API thread bind()s and listen()s
on the API socket.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>