Raise Error when spawning the microvm failed
We use newpid_cloner.c program to clone the jailer. The problem is
that we don't actually get the status code returned by the jailer
when started with newpid_cloner and we always return 0 if cloning
was successful. Using waitpid in new_cloner.c makes the CI system to
hang. Since we have no mean of getting the actual return code of the
jailer process we have to use a hacky alternative and can check that
stderr is empty.

Signed-off-by: Andreea Florescu <fandree@amazon.com>