kvm: added open method with close_on_exec param
Added open method to Kvm. This is now used in Kvm::new() and helps in
avoiding duplicated code. Before we were opening /dev/kvm in the
jailer using libc::open and doing all the checks on the fd file.
Instead we can use the open method from Kvm.

When we are running Firecracker without the jailer we do not expect
for the process to be forked, so /dev/kvm can be open with O_CLOEXEC.
Having this new parameter also helps in migrating the kvm crate to
rust-vmm as in CrosVM /dev/kvm is open with 0_CLOEXEC by default.

Signed-off-by: Andreea Florescu <fandree@amazon.com>