kvm: remove cpuid field from VmFd structure
The VmFd should be generic and having the cpuid as a field is
Firecracker specific. When we are setting the cpuid for each vcpu we
start from the supported cpuid and alter it to match a specified
template. As an optimization, we save the cpuid entries in VmFd and
clone it for each vcpu. This can also be achieved by saving the cpuid
in the Vm structure define in vstate.rs as this structure is specific
to the Firecracker microvm.

Signed-off-by: Andreea Florescu <fandree@amazon.com>