vmm: fix vCPU-thread mapping
Vcpus have an `id` which is a 0-based index, and this `id` matches
the guest vCPU index. We also add this `id` in the name of each
Firecracker vCPU thread so that guest vCPU index can be matched
to a host thread.

This commit fixes this mapping since `.pop()` effectively reverses
the order in which vcpus are consumed from a vec, by also reversing
the order in which we iterate over the `id` numbers.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>