vmm/lib.rs: api_event fd changes
Removed the api_event field from struct Vmm. The Vmm::new constructor
now expects a control_fd: &AsRawFd parameter instead of the old
api_event_fd: EventFd. The new parameter is only used to register an
event source with the epoll set. Since it's now possible to handle
control events from outside the Vmm implementation, the outer logic also
handles the lifetime of the actual object that is associated with the
epoll fd. The from_api member continues to be required, until we
figure out what to do with the API server going forward.

Signed-off-by: Alexandru Agache <aagch@amazon.com>