vmm: added kvm param to Vmm::new
Added a new parameter to the new method of struct Vmm. Previously, the
Vmm would attempt to create a new Kvm object itself (which involved
opening /dev/kvm) via the KvmContext::new constructor. This constructor
has been replaced with KvmContext::with_kvm (also expecting a Kvm param)
which gets passed a Kvm object in Kvm::new. This is a step further
towards the modularization of the Vmm struct, as we no longer impose a
particular way of creating the Kvm object within.

Signed-off-by: Alexandru Agache <aagch@amazon.com>