virtio generic device moves out ownership of Memory
The generic part of a virtio device only needs access to the guest
memory to configure said device.
Once configuration is complete and device is activated, ownership of
the memory should move (instead of clone) to the device implementation.

Also our device does not implement a proper `reset()` so repair the
device-reset unit test.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>