vmm: decouple legacy dev init from IRQCHIP creation
Register legacy devices IRQs as part of legacy device registration,
do IRQCHIP and PIT creation independently as part of (kvm) VM init.

Setting up an interrupt controller should, well.. set up an interrupt
controller (irqchip) without having a dependency on a particular list
of devices that will eventually use it.

Move the registering of legacy devs irqs to `attach_legacy_devices()`
which already does the registering of legacy devs bus addrs.

Before this patch the implementation was not just spaghetti but also
incorrect because it would register irqs for devices not yet attached.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>