remove cpuid template unit tests
The cpuid template unit tests were checking the bytes of each leaf and
registers modified with set_cpuid_entries. Each operation on a register
was duplicated in the tests which makes the unit tests hard to maintain.

A better way to test if the appropriate features are disabled is using
integration tests and check the presence/absance of certain features.
The integration tests will be added in a future commit.

Signed-off-by: Andreea Florescu <fandree@amazon.com>