vmm: run seccomp test in a separate thread
This is needed because we run all tests in a single thread.
Other unit tests will fail if they call a syscall which is not
whitelisted. This is a side effect and we should not have
side effects in unit tests.

Also, since all tests run in the same thread only one of the
tests in default_syscalls/mod.rs will pass (whichever is executed
first). That is because we do not allow SYS_prctl and additional
filters cannot be added.

For further details, please check the prctl defintion:
http://man7.org/linux/man-pages/man2/prctl.2.html

Signed-off-by: Andreea Florescu <fandree@amazon.com>