vmm: run test_signal_handler in separate thread
We need to spawn a new thread when we test register_signal_handlers
because to test that function we install syscall filters. Since all
tests run in a single process the syscall filters will be enabled
for all unit tests that run after `test_signal_handler`. This side
effect causes failure in other tests.

Signed-off-by: Andreea Florescu <fandree@amazon.com>