src: fix main.rs test
The test was not deleting /tmp/firecracker.socket in case one of
the asserts failed. Instead of doing an assert, remove the file
before running the test.

Also, increased FIRECRACKER_INIT_TIMEOUT_MILLIS to 150ms. Before the
value was 100 which was causing the tests to fail when running under
kcov with Rust-1.35.0.

Signed-off-by: Andreea Florescu <fandree@amazon.com>