devices/vsock: connection state machine
Added the vsock connection state machine implementation. This resides in
a the vsock `csm` submodule, that exports a `VsockConnection` struct,
which can handle the vsock traffic between a guest-side AF_VSOCK socket
and a generic host-side `Read + Write + AsRawFd` stream.

`VsockConnnection` implements `VsockChannel` for sending and receiving
vsock packets, and `VsockEpollListener` for getting notified about
events available for the backing host-side stream.

The code is documented in depth, with all the additional implementation
details.

Signed-off-by: Dan Horobeanu <dhr@amazon.com>