Refactored Vmm::add_event
The `add_event` function used to take ownership of the file descriptor
of the event it needed to add, and then would return ownership to the
caller via a return. According to the rust style guide, one should
minimize ownership assumptions i.e. with regards to function parameters,
if possible, borrow rather than move. This leads to this refactor, which
makes `add_event` take it's parameter as a reference.