Extend SeccompCondition with arg len parameter
Some libc implementations (for example musl) perform sign extension when
passing an int argument to a linux system call. This can result in
incorrect behavior when a bpf rule treats any syscall arguments as
a 64 bits variable.

We define a way to compare only the least significant 32 bits of a
syscall argument in order to deal with this corner case.

We also add some regression tests

Signed-off-by: Serban Iorga <seriorga@amazon.com>