Remove testing redundancy.
Testing coverage implicitly tests all unit tests (but not doc tests).
This commit makes it so that on x86_64 (where we can run coverage tests)
unit tests are not redundantly tested -- and also runs doc tests
explicitly (since they are not tested by coverage). It does not change
testing behaviour for aarch64.

Signed-off-by: Tamio-Vesa Nakajima <tamiove@amazon.com>