Remove `Tap` device from `NetworkDeviceConfig`.
This commit removes the `Tap` device file from the
`NetworkDeviceConfig`. Now, each time we need this device file, or we
need to check that it can be opened, we will explicitly open it using
`open_tap`. This makes it so that `NetworkDeviceConfig` holds only
configurations, not any "live" data, such as file descriptors.

Signed-off-by: Tamio-Vesa Nakajima <tamiove@amazon.com>