Fix for off-by-one error in checked_offset
When reading an object from a `MemoryRegion`,
we would normally make an out of bound
check, based on the expression:
"obj.addr + obj.len < region_end(region)".

This expression will be resolved to `false`
for every obj which sits at the end of a region,
blocking reading the object from the memory.

The fix consists in changing the expression
like follows:
"obj.addr + obj.len <= region_end(region)".

Signed-off-by: iulianbarbu <iul@amazon.com>