Implemented new ApiServer that doesn't use Hyper
Replaced the existing API Server with a new one which
uses HttpServer instead of using Hyper and Tokio.

This reduces the number of dependencies from 104 to 74.
Using this new ApiServer enables much easier addition of
new API requests, as can be seen in the following commits.

This commit only brings in the ApiServer framework plus processing
MMDS and InstanceInfo requests, the following commits will provide
processing of the other Firecracker URLs.

Signed-off-by: George Pisaltu <gpl@amazon.com>