seccomp: add missing syscall to tests whitelist
In the seccomp tests the action for non-whitelisted syscalls is
SeccompAction::Errno. With the updated version of Rust a new
syscall is called when running the tests (SYS_futex). Because we
are returning an error for non-whitelisted syscalls, the tests
were hanging as the futex would never be executed successfuly.

Added SYS_futex to the syscall list.

Signed-off-by: Andreea Florescu <fandree@amazon.com>