Update fcuvm container version
- Updated Rust version to 1.38.0
- Updated Python version to 3.6.8

For aarch64, rustfmt is not available with the latest Rust. Hence,
the style test is now temporarily disabled until the rustfmt team
creates a build for aarch64 as well.
This is tracked in:
https://github.com/firecracker-microvm/firecracker/issues/1322

Another side effect of the container update is that the coverage
dropped from 86.0 to 85.6. The same code tested with the previous
kcov version was generating a coverage of 86.1 for the code changes
submitted in this PR. Hence, the coverage drop is more related to
the kcov/rust update than to the changed lines of code.

Signed-off-by: Andreea Florescu <fandree@amazon.com>