vmm: add support for pausing/resuming microVM vcpus
Implemented a control interface through which the Vmm (or another
controller component) can communicate in a request/response model
with each Vcpu.

`Vcpu` exposes a `start_threaded()` function which will run the vcpu
emulation on its own thread. The function returns a `VcpuHandle`
which exposes a control interface to be able to send requests/events
and receive responses from the owned vcpu thread.

Using the above, the control thread can now send commands/events to
the vcpu threads.
Implemented support for `Pause`, `Resume` and `Exit` events. These
effectively pause, resume, or respectively stop the microVM by
pausing, unpausing or stopping the vcpu threads.

The vcpus are kicked out of KVM_RUN when the VMM thread needs
to send them events.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>