vcpu: add support for kicking out of KVM_RUN
This is done by making use of the kvm `immediate_exit` flag.

The vcpu is signaled and the `immediate_exit` flag is set in the
signal handler. This leads to race-free interruption of KVM_RUN.

In order to access the `immediate_exit` flag for a particular vcpu
from a global context such as the signal handler, we use the TLS
of that vcpu thread to gain access to the vcpu's data.

This commit tries to keep the flow as safe as possible, and
in that spirit, the following vcpu methods are defined:
- `init_thread_local_data(&mut self)` - associates the `self` vcpu
   to the current thread.
- `reset_thread_local_data(&mut self)` - deassociates the `self`
   vcpu from the current thread.
- `Vcpu::run_on_thread_local<F>(fn: F)` - runs `fn` passing it
   a reference to the vcpu in TLS of local thread. This model
   enforces a limited lifetime (lifetime of `fn`) for the unsafe
   vcpu reference.
- `Vcpu::register_kick_signal_handler()` - Registers a signal
   handler that sets kvm `immediate_exit` using
  `run_on_thread_local()` in order to kick the vcpu running on
   the current thread.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>