simplify read of block requests
Instead of reading fields of the block request one by one and
account for offsets, read all the request header in one go
using a RequestHeader structure and the read_obj functionality
from guest memory.

This simplifies the code because we don't do manual reads anymore
and we don't need the orfan functions such as read_sector and
read_request_type. Parsing a RequestType from an u32 is now
implemented with std::convert::From.

This code only works on little endian platforms.

Signed-off-by: Andreea Florescu <fandree@amazon.com>