serial: refactor the emulation code
Moved the emulation code from the IoBus implementation to the serial
code. This is a preemptive change for when we are going to consume
devices from rust-vmm. The emulation part of the Serial is not related
to the Bus and should sit in the structure where all the other
emulation related functionality is placed. This change enables the same
serial to be used with different Bus implementations and decouples the
transport from the emulation.

Signed-off-by: Andreea Florescu <fandree@amazon.com>