Operations on EventHandler return vec
For the polly usecase it is safe to assume that None and an
empty vector returned by handle_read, handle_write, ... are
equivalent. The code because cleaner when returning an empty
vector instead of None because you don't have to check the return
value or unwrap when calling the functions.

Signed-off-by: Andreea Florescu <fandree@amazon.com>