split controller into resources, builder and controller
The VmmController is unnecessarily complex because it tries to
handle both the creation operations and the runtime operations
of a Vmm. It can be split in a builder module and a runtime
controller object.

The builder pattern doesn't really mix well with our REST(ish) API
for building a microVM. Most requests work with various microVM
resources' data. These requests model best on a data-store-model.
The StartMicroVM request tries to use all the resource data and
boot a Vmm using it, _but_ it should not consume or mutate the data
because any failure in the booting process then leaves the whole
thing in an unusable/undefined state.

To address this issue, let's just create a vm-resources object that
manages the various microVM resources. This can then be safely used
for all API requests working with resource data. By separating the
Vmm and its resources we can manage their possibly independent
lifetimes.

This model should also help with future devices work where we want
to have the device register its resources with the Vmm instead of
the Vmm creating and managing the device.

The microVM lifetime and how resources are owned and managed is now
clearly defined as two stages:
1. There is a pre-boot vmm-building loop where pre-boot-supported
   API requests are handled. These mostly work on the resource
   collection defining the microVM. At this time, Firecracker acts
   as a data-store exposing a REST API for managing the data (microVM
   resources). A successful `StartMicroVm` command (successful boot)
   will move the microVM into the post-boot state, while failures are
   recoverable as they do not dirty/mutate the data.
2. There is a post-boot `VmController` which runs a control loop that
   handles `Vmm` internal events and also handles post-boot-supported
   API requests.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>

builder becomes static functionality

The 'building' part of the Vmm is now a static recipe with
no ownership of the resources comprising the Vmm or of the
Vmm itself.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>