move logger code from controller to logger.rs
We have the logger crate which provides the logger implementation,
we then have vmm/src/vmm_config/logger.rs which provides the
logger configuration (LoggerConfig) as used in our implementation.
Move the code that initializes the logger based on the LoggerConfig,
in the same file where the config is defined. The controller or the
vmm then simply calls the function at the appropriate time.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>