seccomp: use SeccompFilter instead of SeccompLevel
This commit moves all of the logic of dealing with Seccomp Levels in the CLI
instead of having it generated based on number somewhere in the Vmm.

The purpose of this commit is to allow users to control seccomp filtering to
a much higer degree, by enabling custom filters, including separate custom
filters per vCPU, VMM and API server.

Signed-off-by: Petre Eftime <petre.eftime@gmail.com>