surprise! scrap the builder go for a data-store model
The builder pattern doesn't really mix well with our REST(ish) API
for building a microVM. Most requests work with various microVM
resources' data. These requests model best on a data-store-model.
The StartMicroVM request tries to use all the resource data and
boot a Vmm using it, _but_ it should not consume or mutate the data
because any failure in the booting process then leaves the whole
thing in an unusable/undefined state.

To address this issue, let's just create a vm-resources object that
manages the various microVM resources. This can then be safely used
for all API requests working with resource data. By separating the
Vmm and its resources we can manage their possibly independent
lifetimes.

This model should also help with future devices work where we want
to have the device register its resources with the Vmm instead of
the Vmm creating and managing the device.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>