Vmm RPC interface exposed by VmmController
VmmAction was defined in the `api_server` crate while VmmActionError
was defined in the `Vmm`. The two enums actually define a RPC API
interface to the Vmm.

The VmmController (partly) implements that RPC interface to the
underlying Vmm, so make it also define and expose the RPC API
interface definition.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>