no user-facing errors in Vmm
The Vmm should not try and care about user-facing errors.
The Vmm defines its own set of specific errors which will be later
encapsulated and appended with extra information by the upper layers.

Also make a change to the resources store to now own the kernel-file
and cmdline instead of just descriptions of them.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>