builder: remove dependency on VmResources
The builder module should be usable without VmResources,
other users of Vmm might not want to build using a data
store like VmResources, and could want to build using a
different recipe.

Decouple all builder functions from VmResources, except
the main default build_microvm() which is the Firecracker
Vmm building recipe and which is fine to work with
VmResources.

Other flavors of Vmms can be built by creating other similar
recipe functions which call the building functions with
custom data or in a custom order.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>