metrics driven by polly outside of Vmm
Both LOGGER and METRICS are Firecracker specific, not Vmm specific,
and thus should not be tied to the Vmm.
Logger has already been taken out of Vmm, now take the metrics
periodic logging logic out of the Vmm as well.

A PeriodicMetrics object is defined in the firecracker crate, which
uses polly to drive periodic reporting of METRICS.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>