devices: remove RawIOHandler trait
Remove the `RawIOHandler` trait exposed by `bus.rs`. This was
previously used to get a handle on the Serial device to be able
to externally inject raw input in it.

With the new Serial design, that is no longer necessary, the Serial
device itself can be built with any input/output sources and it will
handle input and output on them internally.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>