Add 'InstanceForceStop' action to forcefully stop all VCPUs
In the case that the guest kernel is either non-cooperative (i.e. does
not shut down when Ctrl-Alt-Delete is sent), or hangs and does not
respond, it's convenient to be able to forcefully stop a microVM.

The newly-added 'InstanceForceStop' action accomplishes this by telling
all VCPUs in the microVM to exit, building off the work done in #1414.

Signed-off-by: Andrew Dunham <andrew@du.nham.ca>