Simplified the interface of EventManager
Removed the pollable.rs abstractions.

Removed the wrapper over the HashMap of (fd, callback). We now use
just insert, contains_key and get. Get does not create a clone of
the entry anymore as it is no longer needed.

Removed EventHandlerData. The EventManager only keeps track of
pollable fds and their corresponding callback.

Removed the implementation of cascaded epoll as it is not needed.

Removed the PollableOp builder. For registering events, the subscriber
returns the epoll_events in a vector. The other usecase for PollableOp
are now covered by calling the corresponding operations directly on
the event manager mutable reference.

The EventHandler (now called Subscriber) trait now only has 2 methods:
- interest_list: returns a list of events that the subscriber is
  interested in. These are added to epoll as soon as add_subscriber()
  is called.
- process: handles the processing of the event. This is replacing
  handle_* functions from the previous interface in an attempt to
  give some more flexibility to the subscriber in terms of order in
  which events are processed, but also handling of errors which are
  now passed to the Subscriber implementation.

Signed-off-by: Andreea Florescu <fandree@amazon.com>