Added a way for updating the monitored events
The implementors of Subscriber should be able to update the
events they're interested in. For this to be possible, we need
a method that returns a clone of the existing Subscriber.

We also pass a mutable reference to the event_manager when
processing events so that the subscriber can call ADD/REMOVE/MODIFY
directly.

Suggested-by: Dan Horobeanu <dhr@amazon.com>
Signed-off-by: Andreea Florescu <fandree@amazon.com>