decouple Vmm from io::Stdin
Created `vmm::VmmEventsObserver` trait that defines two functions which
will be called on microVm boot and on microVm teardown respectivelly.

The Vmm now holds an Option<Box<dyn VmmEventsObserver>> and calls the init
and stop functions implemented by the actual custom object when the vmm
does `start_vcpus()` and `stop()` respectivelly.

The Firecracker-specific Vmm is given a `SerialStdin` which implements the
`VmmEventsObserver` trait and which sets the terminal in raw mode on boot
and back to canonic mode on Vmm stop.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>