Fixes for the arg parser
App struct was renamed ArgParser and app specific fields
were removed in order to keep in `arg_parser.rs` only the
functionality that is related to a command line argument
parser.
ArgInfo struct was renamed Argument because it contains
the `user_value` field which is not actually an info about
the argument.

Signed-off-by: Laura Loghin <lauralg@amazon.com>