vmm: Make output of serial console configurable at Vmm level
Currently, PortIoDeviceManager decides that the output of the serial should go
to stdout(), this is not flexible at all for any consumers of the Vmm struct
which may want to redirect the output to any object that implements Write +
Sync.

Fix this by just adding a new extra parameter to Vmm::new and plumb everything
all the way to the PortIODeviceManager.

Signed-off-by: Alexandru Gheorghe <aggh@amazon.com>