Versioned serializer/deserializer support.
This commit introduces 2 new crates:
- snapshot
- snapshot-derive

The "snapshot" crate defines a simple API to save/load
snapshots from files/memory slices/or anything that
implements  both Read and Write traits. It also exports
the Versionize trait which defines an abstract interface
for dealing with versions while serializing.

The "snapshot-derive" crate exports the "Snapshot" derive
macro which generates code for the Versionize trait implementation.
Users of this procedural macro must also decorate the structures/
enums/unions with versioning information as showcased in the unit
tests (snapshot/src/lib.rs).

This implementation is based on a design which defines
2 translation layers: field level and semantic translators.
The field level translator decides if a field is or is not
serialized/deserialized to/from a target/source version.
Field translators can be generated using the procedural macro.
The semantic translator will handle the semantic changes
that occur when adding or removing fields and requires
manual implementation.

Currently versioned serialization support is limited to:
- primitives: signed/unsigned integers(8/16/32/64 bits), arrays, vecs and strings.
- simple enums with specified u16 discriminants.
- structs of primitive fields or other types implementing Versionize
- unions of primitive fields or other types implementing Versionize

Signed-off-by: Andrei Sandu <sandreim@amazon.com>