tests: set logger for every spawned microvm
Configured the logger in every integration test that spawns a
microvm in order to always check that the Firecracker process
has indeed started. The logger is configured in test_logging.py
still via API requests and in the other tests via command line
parameter. We create a new thread that redirects the console
output to the logging file for a better monitoring of the errors
that may appear.

Signed-off-by: Laura Loghin <lauralg@amazon.com>