pvh: Control PVH capabilities using optional feature
Enable the optional "pvh" feature to build a Firecracker binary
that is able to boot Linux guests using the PVH boot protocol.
Build a firecracker binary with the PVH capability by running:

tools/devtool build [--release] -- --features pvh

and the resulting Firecracker binary will use the PVH boot
protocol as the preferred choice for booting a guest, if the
guest kernel supports it.

Signed-off-by: Alejandro Jimenez <alejandro.j.jimenez@oracle.com>
Acked-by: Liam Merwick <liam.merwick@oracle.com>