Add metrics for vsock virtio device
Add metrics tracking for the vsock device similar to the network virtio
device.

This commit also adds unit tests for the some of the vsock metrics.
Since not all of the metrics can be tested in a predictable manner (such
as read/write fails), we only test the fields that can be properly
controlled within an unit test.

Signed-off-by: Joshua Abraham <sinisterpatrician@gmail.com>
Signed-off-by: Gabriel Ionescu <gbi@amazon.com>