Alphabetize API definitions
Place definitions in alphabetical order.
This aids new and existing users to quickly scroll to the approximate
location of a definition.
Without this order each definition needs to be scanned.

Place `properties` and `required` items in alphabetical order.
This aids new and existing users to visually scan to a item.
Without this order each item needs to be read.

Signed-off-by: Begley Brothers Inc <begleybrothers@gmail.com>