test: rate limiter: avoid false negatives
We have short iperf3 tests to measure TCP throughput when using
network rate-limiters. When on of these tests reports a failure,
re-run it with a longer duration to avoid false negatives.

Real device rate-limiting problems should be exacerbated by longer
tests which means this mechanism should not provide false positives.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>