devices: serial: create with receive irqs enabled
Create serial devices with receive interrupts enabled when
a valid output device is attached to them.

This allows to directly use the serial even when loading a
microvm from snapshot without having to (de)serialize the
serial console (as long as all the events and interrupts are
properly set up).

Signed-off-by: Adrian Catangiu <acatan@amazon.com>