vmm: add metrics for create, load, pause, resume
Add performance metrics that measure the duration for
pausing, resuming the microVM, creating a full or diff
snapshot and loading from a snapshot, from the VMM
perspective.
If there are more than one `/snapshot/create` request
in a minute (until the metrics are flushed), only the
duration of the last snapshot creation is stored in the
metric. If the user is interested in all the durations,
a `FlushMetrics` request should be sent after each
`create` request.

Signed-off-by: Laura Loghin <lauralg@amazon.com>