update to vm-memory 0.2.1
In rust-vmm/vm-memory 0.2.0, the functions read_obj and write_obj
are not doing atomic accesses for all combinations of platform and
libc implementations. This is caused by everything being written
at a one byte granularity. The reads and writes eventually translate
to memcpy, which in some cases are not optimized to write at a higher
granularity. Since accesses are not guaranteed to be atomic, using
vm-memory in the virtio implementation causes undefined behavior
because the requirement of 2-byte aligned accesses of descriptor
indexes cannot be fulfilled.

The patch that fixes this issue is included in the 0.2.1 vm-memory
update.

There were no Firecracker releases after 2020-03-03; all existing
releases, including those used by AWS Lambda and AWS Fargate, are
not affected.

Related to: https://github.com/rust-vmm/vm-memory/issues/93

Signed-off-by: Andreea Florescu <fandree@amazon.com>