Use KVM_GET_CPUID2 in Vcpu::save_state().
Currently the CPUID is cached *only* at microvm boot time. The restore
path fails to cache the CPUID from the snapshot and a bogus value will
be saved when a new snapshot is created from the current state. When such
a snapshot is restored it will cause the in-kernel KVM LAPIC emulation
to setup incorrectly and other undefined behaviour.

This commit removes CPUID caching from Vcpu and calls KVM_GET_CPUID2 on
the Vcpu::save_state() path.

Signed-off-by: Andrei Sandu <sandreim@amazon.com>