Permission: Run tests as root with all cap. on
Closes issue #1980.

While Docker is often used with root access to all capabilities,
this can be a surprise to users coming from podman.

We don't immediately know the minimal set of capabilities the
root user needs to run the Firecracker test suite.

Until then the best we can do is:
1. Tell the user what is about to happen.
2. Get user agreement.
3. Log what they have agreed to.

The logging is generally useful. However it is primarily intended
to inform users who may have inadvertently done something like:

   $(yes | ./tools/devtool test)

This is a workaround until a we identify the minimal set of root
capabilities needed to run the Firecracker test suite.

Signed-off-by: Begley Brothers Inc <begleybrothers@gmail.com>