ugly partial fix for content-length check
If the user was sending a request with a content-length
value > request body size, firecracker was hanging waiting
for other input to achieve the target value.
`self.body_bytes_to_read` is set, when reading a request,
either with the content-length from it (if user passed one)
or with the request body size. If content-length is set by
user with a bigger value than the expected one, the process
hangs because it doesn't check for EOF, but waits until
the target value is achieved. Added a new field in
`HttpConnection` that counts the number of bytes read from
body. If at the current try_read() iteration the number of
bytes read is < BUFFER_SIZE (1024B) then it either means we
are reading the first part of the body or the last one of it
(currently micro_http reads the request in chunks of 1024B).
In the second situation, if the number of total bytes read
is != expected content-length, it means we passed an invalid
length. This solution doesn't work for request body sizes =
multiple of 1024B. For an actual fix, a non-intrusive fix
doesn't seem to be possible :(.

Signed-off-by: Laura Loghin <lauralg@amazon.com>