[virtio][net] simplify Net::interrupt_evt checks
In the unit tests we have to check that the right amount of writes have
been performed on Net::interrupt_evt. We are doing this in 2 steps:
1. write(1) in order to make sure that read() won't block
2. read() and check the result
We create a function for this in order to avoid duplicated code.

Another solution would be to open Net::interrupt_evt with O_NONBLOCK in
the tests but it would rather add complexity.