dev_mgr: adjust legacy registration on aarch64
Legacy devices on aarch64 are mmio devices. We need
to be able to restore MMIODeviceInfo for these devices when
loading microVM from snapshot. As a result, we need that the
registration functions for these device to allow a parameter
that tells them if they need to allocate or reuse a certain slot.

Signed-off-by: Diana Popa <dpopa@amazon.com>