tests: add 'nonci' pytest marker
Added 'nonci' pytest marker which marks a test as a nonci test.
These tests will be deselected if they are not selected by a mark
expression (e.g -m nonci). One can not run both ci tests and nonci
tests in one pytest session because providing a mark expression will
deselect all tests which does not conform with the expression.

Signed-off-by: Iulian Barbu <iul@amazon.com>