tests: default numa_node set to None
When using confinement for the docker dev container, where tests will
run, we will usually restrict the cpulist available to the jailer.
The integration tests assumes that the jailer is going to use
always the NUMA node 0, if not explicitly told otherwise. This conflicts
with the confinement, since it might restrict the jailer to a subset cpulist
of the NUMA node 0. Changed the default numa node to None to avoid this.

Signed-off-by: Iulian Barbu <iul@amazon.com>