GDB server: Alter address-translation approach
Removed the option of setting breakpoints at linear addresses in
early boot. The support for it contained a work-around and not
all inconsistencies at the client could have been solved on the
server side (e.g.: "info breakpoints" would not display such a
breakpoint as "hit" when it was the case).

Signed-off-by: Marius-Cristian Baciu <baciumar@amazon.com>