Temporarily disable the cargo audit test.
The security db (https://github.com/RustSec/cargo-audit)
has broken compatiblity with cargo audit v10. A fixed cargo
audit version v12 is available, but  it requires an upgrade of
the rust toolchain on top of the current release. This commit
is a temporary fix until we unshare toolchain dependency
of Firecracker and cargo audit binary.

Signed-off-by: Andrei Sandu <sandreim@amazon.com>