Add support for pausing and resuming emulation
When pausing or resuming the VM, alongside vcpus device emulation is
now also paused or resumed.

For reduced complexity and runtime overhead, this change only affects
virtio devices. Legacy devices are controlled by vcpus (which get
paused) and use no resources external to firecracker so their state
remains consistent even if they're not explicitly paused.

Implementation details:

To keep an unified model for both fresh booting a VM and restoring
one from snapshot, we follow the same pattern with devices as with
vcpus: devices start off in the paused state and they are resumed
alongside vcpus.

Device events registration is moved from device attachment to Vmm,
to device resume path.

The MMIODeviceManager was enhanced with a list of subscribers to be
able to directly access the Subscriber trait functions on the devices
without having to downcast BusDevices all the way down to their
concrete types then upcast back to Arc<Mutex<dyn Subscriber>>.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>