virtio: add option to enable flush
If a device advertises VIRTIO_BLK_F_FLUSH, then it must sync
the data to physical media. The current flush implemention
only flushes cached data to the host page cache, which
can lead to incoherency between VM and physical media.
Unfortunately, during boot there can be up to six flush
requests on a read-write root partition, which can slow
down booting. This means that the current paradigm (of not
flushing data to physical media) must remain the default.

This commit adds a drive option 'want_flush' that allows each
drive to specify whether they require consistency and are willing
to take the performance hit of syncing data out to media. The
default is the existing behaviour, namely no flush.

For drives that require the flush, the virtio layer will first
flush any cached data (the current behaviour) and then force
the pages in the host page cache for the drive out to physical
media via sync_all().

This has been verified by running a 'fio' test with
sync_on_close in the VM and verifying via 'perf' that fsync()
on the host is actually invoked. Also, the bandwidth in the
VM with flush enabled is also now in line with the underlying
hardware.

The invocation of sync_all() has been refactored to be in accordance
with other function invocations in the request handler.

This change also requires that the sys_fsync system call is in the
list of approved seccomp calls.

Signed-off-by: Peter Lawthers peter.lawthers@esrlabs.com