virtio: add vm level flush tests
This adds python tests to verify the correct operation
of flush. The features file in /sys/block/<drive>/device/features
is checked to ensure that it is correct with respect to the
optional 'want_flush' drive setting. Metrics also verify that
by default, no flushes are received. Only when 'want_flush' is
true should the 'flush_count' block metric increase.

Signed-off-by: Peter Lawthers peter.lawthers@esrlabs.com