tests: added support for _exec without async
`_exec` depends on asyncio runtime which does not
fit well when used within multiple threads context,
since an asyncio loop is dependent on a single thread.
Support for executor and asyncio can be added to the
testing framework, but it does not bring significant
value over the current solution. Added an utility function,
`run_cmd_sync` to cover this lack of interface.

Signed-off-by: Iulian Barbu <iul@amazon.com>