metrics: introduce a new type of metrics
With the passing of the time, in Firecracker started
to shape out 2 types of metrics:
* the types of metrics needed for counting purposes -> they
are represented by 2 values, one of them is the active counter while
the other one is the value that was last flushed. With each flush, we would
 compute the delta between them
* the types of metrics needed for storing persistent values of a specific
measurement, for example the ones used for obtaining the process start up
time. This type of metric does not need to keep around a second value
since the value being flushed is always the one the system updates.

This approach also rids us of any memory reordering that may happen on weak
memory model.

Signed-off-by: Diana Popa <dpopa@amazon.com>