balloon: fix intermittent failures in arm tests
There is an intermittent failure in the balloon integration tests when
running them on arm instances. This may be due to the differences in
the way the kernel consumes memory at boot and the fact that the
balloon inflate operations are slower on arm than on x86. This commit
addresses this problem.

Signed-off-by: George Pisaltu <gpl@amazon.com>