rpc_interface: tests: add mock Vmm
The RPC interface code has the purpose of transforming locally
defined RPC commands (VmmActions) to programmatic API calls to
the underlying components.

In order to efficiently test the RPC interface code, we could create
mock objects for these underlying components and make use of these
mock objects to control and validate expected RPC interface behavior.

This commit adds `struct MockVmm` which, when under config(test), is
exported as `Vmm`.

With a mock Vmm struct we can easily test the RuntimeApiController,
as well as the the transition between the two stages (preboot/runtime).

Signed-off-by: Adrian Catangiu <acatan@amazon.com>