balloon: fix reinflate intermittent test fail
There is an intermittent failure in the balloon reinflate integration
test on arm instances. This may be due to the differences in the way
the kernel consumes memory at boot and the fact that the balloon
inflate operations are slower on arm than on x86. There seems to be a
higher amount of memory freed after boot by the kernel, and by working
with 256 MB machines and checking for small differences in memory
after inflates, 10 MB, sometimes when dirtying memory during the test,
some of that memory used at boot was being uses instead of demand
paging new memory, which would skew the test results.

Signed-off-by: George Pisaltu <gpl@amazon.com>