arg_parser: simplify for flag-type arguments
This commit changes `Value::Bool(bool)` to `Value::Flag` since the value
only encodes whether an argument with no values (also referred to as a flag)
is present or not.
The inner bool is redundant (since it is always true) and confusing (because
it suggests it could also be false).

Also do a quick rename of `Value::String` and `Value::Vector` to
`Value::Single` and `Value::Multiple` to avoid confusion caused by the name
collision with the standard types.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>