CI: Refactor sleep calls in balloon tests
Currently, tests in test_balloon.py use sleep() calls with hard coded
values to wait until the RSS of the microvm reaches a stable value
for reading.

The behaviour is changed and the test now waits until the percentage
change in RSS falls under a specified threshold. Right now 0.5% is
used as the threshold.

Signed-off-by: Piyush Jain <piyushkj@cs.utexas.edu>