tests: added limit for maximum opened connections
The micro-http server holds a HashMap of all open connections and
has a limit for the number of active connections at the same time.
The `requests_unixsocket` python package, used by the testing
framework for issuing API requests, opens a new connection for every
API call and keeps it alive even after receiving the response from
the server. These sockets are kept inside the connections HashMap
and might very easily end up in exceeding the micro-http's limit,
triggering a '503' response from the server.
To prevent this from happening, the `pool_connections` argument was
set to 10, consistent with the maximum number of open connections
allowed. Attempts to issue a new API request, after this limit has
been reached, implies closing and removing the least recently used
open fd and then adding a newly opened connection to the pool.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>