CI: adjust boot time initrd on x86_64
The test in question is run with 4.9 kernel, we want
to run the boot time test with the 4.14.174 kernel config
offered in the repo. The boot time obtained with the repo config
is slightly around 190000ms. This difference comes mainly from the
security config that we enable (i.e KPTI).

Signed-off-by: Diana Popa <dpopa@amazon.com>