Forbid snapshots for v0.23 with many devices
Firecracker v0.24 now supports 24 IRQ lines, so it accepts more
devices to be attached. Creating a snapshot with a previous version
(for example: `v0.23.0`) is only allowed if the number of devices
added is consistent to the maximum supported by the requested version.
Otherwise, snapshotting will fail with a dedicated error.
Integration tests were added to validate the snapshot sanity check.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>