use upstream kvm-ioctls
This should reduce the burden of maintaining forks of both kvm-ioctls
and kvm-bindings, and allows us to use the upsream kvm-ioctls while only
replacing the kvm-bindings dependency with one that supports versionize.

As part of this switch, an update to the tests was also needed because
the KVM::new_with_fd_number was removed. The same functionality is
offered through the from_raw_fd method.

Signed-off-by: Andreea Florescu <fandree@amazon.com>