aarch64: setup cache topology
A new module was created for reading the
cache topology information from the host
and parsing it so that it can be appended
to the device tree.
The cache information is retrieved from the sysfs
of the host. Unfortunately, the kernel does not
offer the interface for retrieving the necessary
registers that hold cache information (i.e CSSELR_EL1).

Signed-off-by: Diana Popa <dpopa@amazon.com>