jailer: copy cache info into chroot for aarch64
In order to setup the cache topology we need access
to the sysfs (i.e /sys/devices/system/cpu/cpuX/cache) cause
unfortunately the kernel does not offer the interface for
reading cache info (i.e CSSELR_ELI register which is essential
for doing that is not exposed).
However, once the jailer starts there is no access to the sysfs and
this PR helps with copying the minimal number of info which is needed
for setting up cache topology through FDT on aarch64.

Signed-off-by: Diana Popa <dpopa@amazon.com>