serial: handle EWOULDBLOCK error.
When the serial out buffer is full, and opened in O_NONBLOCK mode,
the guest OS will be blocked waiting for a consumer to read from
the buffer. This is because the serial device does not send an
interrupt, on error. The interrupt is now also triggered in some error
scenarios for writes.

Any new bytes will be lost, until the buffer is freed. The device also
logs these errors and increments the `uart.error_count` metric for each
lost byte.

Signed-off-by: Andrei Sandu <sandreim@amazon.com>
Signed-off-by: alindima <alindima@amazon.com>