block: fix off-by-one error in addr validation
The block device validated addresses from virtio descriptors before
trying to execute requests. In this validation, the `checked_offset`
function of guest memory was used to determine if the slice defined
by the sum of the address and length of the virtio descriptor was
within the guest memory bounds. However, this sum is greater than
the last valid offset, `addr + len - 1`, by 1. This made the block
device mark descriptors with slices at the very end of a region as
invalid, making the last byte of a memory region unusable by the
block device.

This check was performed as the previous guest memory model did not
have it built in. This commit removes the checks as the current guest
memory model validates the addresses before operations. Also added
regression tests for this case.

Signed-off-by: George Pisaltu <gpl@amazon.com>