Fix boottime tests
Renamed the boottime microvm fixtures, so that the
tests use the old PIO-based boot time notification on x86.

Also created new folders for these fixtures in S3:
old_minimal and old_minimal_with_initrd.

This had to be done because the fixtures were pointing to
artifacts that are used in mainline, with the new MMIO boot-timer
device.

Signed-off-by: alindima <alindima@amazon.com>