tests: fix snapshot tests for older versions
Previously, the snapshot tests that were supposed to retrieve
the older Firecracker binaries from S3 were not filtering the artifacts
by version number.

This results in test failures for patch releases, where the test
will use both older and newer binaries.

Added support in the testing framework for this and fixed the
respective tests.

Signed-off-by: alindima <alindima@amazon.com>