devtool: allow fetching private git repos
`devtool` runs the `cargo build` command from inside the container.
In order to enable git authentication for fetching private git
repositories, the public and private SSH key files on the host must
be mapped to the guest. A new `--ssh-keys` flag was added to the
`devtool`'s build functionality to provide a way to forward the SSH
keys to the guest. This functionality is useful when working with
dependencies on libraries which reside inside private repositories.

Signed-off-by: Luminita Voicu <lumivo@amazon.com>