aarch64: relax cache topology restrictions
If some of the files that we look for while constructing
the cache topology for aarch64 are missing, we would return an
error that would block the spawning of the microVM. As per
issue #2407, there are cases where the cache topology of
the host miss some of the cache attributes (for eg the size).
We changed the flow to not throw an error if some of the files
are found and to try to create as much as possible of the cache topology
without blocking on errors.

Signed-off-by: Diana Popa <dpopa@amazon.com>