tests: minor test_signals.py fixes
The generic signal handler integ test now also verifies
that the Firecracker process was indeed terminated.

The test_handled_signals was confusing since it appeared
to be signaling SIGRTMIN+1, whereas in Firecracker, we use
SIGRTMIN for kicking vcpus.
This was actually a confusion in the code, since the SIGRTMIN
value is different across libc implementations. MUSL's sigrtmin
is 34, whereas glibc has 35 (this is the one reported by python
as well).

Replaced the SIGRTMIN in the test with the constant 35 to
avoid confusion.

Signed-off-by: alindima <alindima@amazon.com>