tests: fix imports and name clash with stdlib
During an attempt to add a new file to the framework, the imports in the
statistics module started failing. This was caused by the fact that the
new file name was starting with "test_".

The problem is that since the files here are used and designed as a module,
imports should be relative to this module.

Another issue was that it caused a name clash with the stdlib statistics
module, thus causing another type of import issues.

This commit fixes both of these problems.

Signed-off-by: Gabriel Ionescu <gbi@amazon.com>