Add signalfd signal handling via event manager
Until now, the signal handlers were not async-signal-safe,
since they were using potentially locked resources, for
logging and metrics.

This could lead to thread deadlock, if the thread that is
interrupted to handle the signal was already holding the
logging/metrics locks.

The option of removing the metrics and logging are not viable,
since we need this information for debugging purposes.

In this commit, we replace the classic event handler with a
signalfd. The signalfd is similar to an eventfd, but carries
additional information about the signal that was raised (signo,
code, etc.).

This signalfd is added to the EventManager so that a subscriber
can listen for any events and handle them asyncronously, in
regular userspace context. This lets us use locked resources,
since we are not interrupting the thread from doing its regular
work. We handle the signals when we can, instead of interrupring
the current thread.

For this to work, we also install signal masks to block the
handled signals on the api and vcpu threads and to block the
default handlers on the vmm thread.

One difference from the previous implementation is that, if
the process receives a signal before the eventmanager loop
is run (and before we register the signalfd subscriber),
the default handlers will be used.

Signed-off-by: alindima <alindima@amazon.com>