tests: add SIGSYS integration test
Previously, there was no integration test for SIGSYS
handling. Other handled signals were covered but not
this one.

Added it to the test_generic_signal_handler.

Also added an error log for the case where the FC process
receives a SIGSYS for reasons other than a bad syscall.
(Such as the scenario in this integration test, directly
calling kill).

Signed-off-by: alindima <alindima@amazon.com>