Update build script to watch for seccompiler changes
Seccompiler is run during the build process, being
triggered by the vmm/src/build.rs. However, was
build script is not triggered by a modification in the
seccompiler source code. It was only triggered by a modification
in the target-specific filter file.

Before this commit, if one does a modification in the seccompiler
code, and then builds Firecracker normally, via devtool build
or cargo build, the Firecracker build script would not run.

During development or when building binaries for a release, this
is very inconvenient, because we actually want to test/use
the most up-to-date seccompiler binary.

This is a side effect of the design decision that we made
in order to decouple the seccompiler binary from the firecracker one.

In order to make this work, I added every source file that
seccompiler has, to the watchlist that triggers a build script
re-run.

This means that now, every source code change or JSON file change
for the current target, will re-run the build script and use the
latest seccompiler binary.

This is not the cleanest way of doing this, but there is no other
better way I could come up with. Once seccompiler will be moved to a
separate repository/project, we would get rid of this problem and hack,
because the source code will be kept somewhere else and a manual build
would be neccessary anyway.

Signed-off-by: alindima <alindima@amazon.com>