Glibc SIGSYS patch for sandboxed localtime_r()
The LocalTime::now() function is used by the logger for some
messages (such as warnings).  If the first case of calling
a log message that included the time happened in a sandboxed
thread, glibc would potentially read from the host filesystem
on that thread in response to the log message.

This effect arises from the implementation of how localtime_r()
obtains time zone information if the "TZ" environment variable
is not set.  A function called __tzfile_read() is called:

  https://sources.debian.org/src/glibc/2.28-10/time/tzset.c/#L584

Which triggers an fopen():

  https://sources.debian.org/src/glibc/2.28-10/time/tzfile.c/#L165

Because this information is cached, it appears possible to work
around the problem by making an early call to LocalTime::now()
during initialization.  Then if calls pertaining to logging
come later on the seccomp filtered thread the caching is done
already and problems don't occur.

This permits the alpine demo from the quick start to appear to
run with glibc instead of musl, and built with the standard Rust
on an Ubuntu 20.04 installation (inside of VirtualBox 6.1.20).

Signed-off-by: Cerul Alain <ae@metaeducation.com>