Unwind main() for clean shutdown (vs. libc::_exit())
The design of the existing main() function was such that Firecracker
never unwound the stack to the top level in order to end the program.
Instead, it would run some Firecracker-specific shutdown code and
then call `unsafe { libc::_exit() }`.

That skips some Rust-specific shutdown done by std::process::exit().
But even using that does not call destructors for the stack objects,
so it is not a "clean shutdown":

  https://doc.rust-lang.org/std/process/fn.exit.html

While a "dirty shutdown" can be a bit faster, the benefit of having
non-panic shutdowns run their finalization is that it helps with the
accounting of Valgrind and sanitizers.  It's a good sanity check to
make sure the program could exit cleanly if it wanted to.

This commit attempts a somewhat minimally-invasive means of bubbling
up an "ExitCode".  Exiting was managed uniquely by the Vmm event
subscriber during the process() callback, and requires running the
vmm.stop() method.  So this extends the Subscriber trait with an
extended form of callback named process_exitable() that returns
an optional exit code.

The event loop doesn't have a clear (to me) way to get at the VMM
to call stop() if a non-Vmm subscriber requests an exit.  So the
interface means any other subscriber that overrode process_exitable()
would miss the required stop code.

So it suggests the need for a redesign by someone more familiar with
the architecture.  However, it does accomplish a proof of concept by
getting closer to a clean bill of health from a Valgrind analysis.