Exit cleanly in debug build, abruptly in release build
In order to mitigate the risks of deadlock during shutdown in
production builds, this adds back the behavior of exiting the
process abruptly after writing metrics files.  This also offers
better performance by letting the OS deal with cleaning up
memory and resources.

This limits the behavior of unwinding all the way up to main()
and joining the threads to debug builds.  The policy could be
replaced by an independent flag (if someone thought it important
to run Valgrind on release builds, e.g. because it was too slow
using it with debug ones).