tests: use ramdisk for snapshot perf tests
To provide a safe and consistent CreateSnapshot operation,
Firecracker now flushes all snapshot related host-file as part
of the create snapshot operation.

This operation adds backing-medium specific latency to the
CreateSnapshot operation. To eliminate this variability and measure
firecracker-specific perfomance, use ramdisks in snapshot performance
tests.

Even so, the flushing does add a bit of latency on snapshot creation.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>