seccomp: apply() no longer needs filter ownership
The BPF filter passed to the SeccompFilter::apply function
does not need to get ownership of the filter in order to
pass it to the kernel. It needs to get a const pointer to the
filter, which can also be retrieved from a reference.

The linux kernel does a `copy_from_user` before installing the
filter, so we are sure that we can deallocate the filter once
the function returns. (This is an assumption we were already
by transferring ownership to the `apply()` method, which was
already calling drop before returning).

Since the kernel does a `copy_from_user`, our filter reference
should remain valid. Having looked over the kernel code, the prctl
syscall receives the filter as a `void __user*` and later passes
it to a function that receives it as a `const char*` and copies it
from user memory to kernel memory.

This means that it is safe to reuse this same filter for all
threads, since the kernel does not mess with our memory.

Having done this optimisation, we can now use the same BPF filter
across all vcpu threads, without having to clone() it, by using
an `Arc<BpfProgram>`.

Signed-off-by: alindima <alindima@amazon.com>