seccomp: add byte_limit to filter deserialization
This byte limit is passed to `bincode` to guard against a potential
memory allocation DOS caused by binary filters that are too large.
This limit can be safely determined since the maximum length of a
BPF filter is 4096 instructions and Firecracker has a finite number
of threads with at most 4-letter names.

Added this optional byte limit in seccompiler's `deserialize_binary`
helper function and set a default of 0.1MB in Firecracker.

Signed-off-by: alindima <alindima@amazon.com>