Shutdown API server thread so it can be joined
We want to cleanly exit the API thread so the firecracker process
can gracefully teardown.

Because the API Server thread is in a loop that does a blocking
wait on a socket, the only way to tell it to stop looping is via
information on that socket.  This uses an internal HTTP PUT signal
to request that shutdown.  Rather than forget the thread handle
and the socket bind location, the main thread now sends the
shutdown request and then waits while the thread joins.

The request is for internal use only, however--because it is
supposed to happen after the Vmm has been cleaned up and its
contents destructed.

A different approach for signaling the thread for this purpose
could be used, e.g. to have the API server thread waiting on
multiple events (the socket or this internal signal).  But since
this is only needed for clean shutdown, it is probably good enough.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>
Co-authored-by: Cerul Alain <ae@metaeducation.com>