Unwind main() for clean shutdown (vs libc::exit())
The design of the existing main() function was such that Firecracker
never unwound the stack to the top level in order to end the program.
Instead, it would run some Firecracker-specific shutdown code and
then call `unsafe { libc::_exit() }`.

That skips some Rust-specific shutdown done by std::process::exit().
But even using that does not call destructors for the stack objects,
so it is not a "clean shutdown":

  https://doc.rust-lang.org/std/process/fn.exit.html

While a "dirty shutdown" is easier, the benefit of having non-panic
shutdowns run their finalization is that it helps with the accounting
of Valgrind and sanitizers.  It's a good sanity check to make sure
the program could exit cleanly if it wanted to.

This commit attempts a minimally-invasive means of bubbling
up an "ExitCode".

Signed-off-by: Adrian Catangiu <acatan@amazon.com>
Co-authored-by: Cerul Alain <ae@metaeducation.com>