Simplify Rust integ tests based on new clean exit
Firecracker's Rust integration tests were very hard to work with
because Firecracker was doing 'libc::exit()' which would also
exit the cargo-test process. As a workaround to that the tests'
payload was running under a separate forked process.
The problem with that was that it offered zero debuggability,
the only thing reported being pass/fail.

Using the new clean exit model, tests have been reworked to scrap
the extra 'libc::fork()' and work in the common testing model
where each test can run on a separate thread in the same process.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>