snaps: remove extraneous logs in resume path
The snapshot resume latency performance test has been failing
sporadically. One of the reasons for these failures is the
non-deterministic latency of logging (and flushing logs).

This commit removes a couple of extraneous log entries when
not relevant (when devices are not active, kicking is noop).
It also sets the log-level in the performance test to 'Error'
instead of 'Info' to minimize logger-induced jitter.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>