balloon: fix stats after snapshot restore
We save the balloon stats descriptor head in a special field in the
balloon device (`stats_desc_index`). When the stats timer expires and
we want new stats, we take the descriptor and place it in the used
ring, signaling the driver that we did so.

The problem is that when saving the device state, while we do save the
descriptor, we do not configure the restored device to use it. This
makes the interaction broken because the descriptor is essentially
lost, and the driver will not send a new one until we place this one
on the used ring.

This means the device will not get updated statistics after a resume
event and there will be a lot of log messages from the device.

This commit fixes the issue by actually restoring `stats_desc_index`
field in the device on a snapshot restore event. Also updated
CHANGELOG.md.

Signed-off-by: George Pisaltu <gpl@amazon.com>