Revert to classic signal handlers from signalfd
In a previous PR, we replaced the classic signal handlers
with an async signalfd mechanism, to prevent deadlocks
on acquiring the metrics and logging mutexes.

This approach turned out to be flawed. Older linux
kernels do not support catching SIGSYS signals on
signalfds.

Moreover, in a multithreaded scenario, if a signal is
specifically directed to another thread than the process
leader, or is not directed to the process as a whole,
those signals would not be caught at all.

This is not ok for signals that are synchronously sent by
the kernel (SIGSYS, SIGSEGV, etc.).

Therefore, in this commit, we roll back to using classic signal
handlers.

Signed-off-by: alindima <alindima@amazon.com>